Separate refresh loop to refresh_loop.py module
